Fix os_neutron and etcd handler naming conflict.
The handler to do a 'systemctl daemon-reload' in both os_neutron
and the etcd dependancy have the same name.  This was causing the
etcd one to run which was being skipped.  The end result was the
neutron services were stuck running old code and never reset on
upgrades.  Renaming the os_neutron handler for reloading the
daemon seems to fix the issue and run the handler from os_neutron
without being skipped as before.

JIRA: RE-1296